main: Print the secret key to stdout if I/O fail
This commit lets curvy output the secret key in Base16 to stdout if it
cannot be written to the specified output file.

The reason behind this is, that after an I/O failure, all of the
computing was rendered useless, therefore outputting the secret key on
stdout will not waste the energy.